+++ src/main/java/calculator/Main.java
@@ -1,6 +1,8 @@
-import calculator.StringCalculator;
-import view.InputView;
-import view.ResultView;
+package calculator;
+
+import calculator.domain.StringCalculator;
+import calculator.view.InputView;
+import calculator.view.ResultView;
 
 public class Main {
     public static void main(String[] args) {
