+++ src/test/java/calculator/OperatorTest.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import calculator.domain.Operator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
