+++ src/test/java/calculator/StringCalculatorTest.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import calculator.domain.StringCalculator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
